Refactor navegación
- Uso switch navigator para manejar correctamente el tema de la pantalla de personal info. Esto soluciona varios bugs de presionar Back button en Android. LoadingAppScreen reemplaza al antiguo Main container.
- Extra: Muevo algunas funciones de envío de datos, que no son acciones de redux, a un archivo aparte (send_data.js).
