+++ src/actions/personal_info_actions.js
@@ -1,22 +1,9 @@
-import {AsyncStorage} from "react-native";
-import {sendPersonalInfo} from "./main_actions";
-
 export const AGREE_TERMS = 'AGREE_TERMS';
 
 export function agreeTerms() {
-    return (dispatch, getState) => {
+    return (dispatch) => {
         dispatch({
             type: AGREE_TERMS,
-        });
-    }
-}
-
-export function savePersonalInfo(personalInfo, afterSaveCallback) {
-    return (dispatch, getState) => {
-        personalInfo['sentToBackend'] = false;
-        AsyncStorage.setItem('@moravec:personalInfo', JSON.stringify(personalInfo)).then(() => {
-            afterSaveCallback();
-            sendPersonalInfo(personalInfo);
         });
     }
 }
